Inline SubtleLabel, TokenAmount. Use TokenBadge.
